More cleanups.  Bug fixes: 1) x2,y2 guaranteed to be filled when making paths 2) PruneDeadends() recurses to check for deadends it creates.
