+++ player.go
@@ -33,10 +33,8 @@ func (p *Player) Attack(m *Monster) string {
 // -----------------------------------------------------------------------
 func (p *Player) InfoString() string {
 	info := fmt.Sprintf(
-		"Level:%2d  Gold: 4       Hp:11(20)  Str:16(16)  Arm: 4   Exp: 2/14      (%d,%d)",
+		"Level:%2d  Gold: 4       Hp:11(20)  Str:16(16)  Arm: 4   Exp: 2/14",
 		p.depth,
-		p.X,
-		p.Y,
 	)
 	return info
 }
